+++ .build/parent-maven-pom.xml
@@ -671,7 +671,7 @@
       <dependency>
         <groupId>org.apache.cassandra</groupId>
         <artifactId>cassandra-all</artifactId>
-        <version>4.1-alpha2-SNAPSHOT</version>
+        <version>@version@</version>
       </dependency>
       <dependency>
         <groupId>io.dropwizard.metrics</groupId>
