+++ .build/sstableloader-deps-maven-pom.xml
@@ -1,0 +1,48 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+  Licensed to the Apache Software Foundation (ASF) under one or more
+  contributor license agreements.  See the NOTICE file distributed with
+  this work for additional information regarding copyright ownership.
+  The ASF licenses this file to You under the Apache License, Version 2.0
+  (the "License"); you may not use this file except in compliance with
+  the License.  You may obtain a copy of the License at
+      http://www.apache.org/licenses/LICENSE-2.0
+  Unless required by applicable law or agreed to in writing, software
+  distributed under the License is distributed on an "AS IS" BASIS,
+  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  See the License for the specific language governing permissions and
+  limitations under the License.
+-->
+<project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd" xmlns="http://maven.apache.org/POM/4.0.0"
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
+  <modelVersion>4.0.0</modelVersion>
+  <parent>
+    <groupId>org.apache.cassandra</groupId>
+    <artifactId>cassandra-parent</artifactId>
+    <version>@version@</version>
+    <relativePath>@final.name@-parent.pom</relativePath>
+  </parent>
+  <artifactId>cassandra-sstableloader</artifactId>
+  <version>@version@</version>
+  <name>Apache Cassandra SSTableLoader</name>
+  <description>Standalone SSTableLoader for Apache Cassandra.</description>
+  <url>https://cassandra.apache.org</url>
+  <inceptionYear>2025</inceptionYear>
+  <licenses>
+    <license>
+      <name>The Apache Software License, Version 2.0</name>
+      <url>https://www.apache.org/licenses/LICENSE-2.0.txt</url>
+    </license>
+  </licenses>
+  <scm>
+    <connection>scm:https://gitbox.apache.org/repos/asf/cassandra.git</connection>
+    <developerConnection>scm:https://gitbox.apache.org/repos/asf/cassandra.git</developerConnection>
+    <url>https://gitbox.apache.org/repos/asf?p=cassandra.git</url>
+  </scm>
+  <dependencies>
+    <dependency>
+      <groupId>org.apache.cassandra</groupId>
+      <artifactId>cassandra-all</artifactId>
+    </dependency>
+  </dependencies>
+</project>
