+++ .circleci/config_template.yml
@@ -63,6 +63,13 @@ default_env_vars: &default_env_vars
     # The number of times that new, modified or manually specified fqltool unit tests should be run.
     REPEATED_UTESTS_FQLTOOL_COUNT: 500
 
+    # Comma-separated list of tests that should be included in the repeated run for sstableloader unit tests,
+    # in addition to automatically detected new and modified tests. For example:
+    # REPEATED_UTESTS_SSTABLELOADER: org.apache.cassandra.tools.LoaderOptionsTest
+    REPEATED_UTESTS_SSTABLELOADER:
+    # The number of times that new, modified or manually specified sstableloader unit tests should be run.
+    REPEATED_UTESTS_SSTABLELOADER_COUNT: 500
+
     # Comma-separated list of tests that should be included in the repeated run for long unit tests,
     # in addition to automatically detected new and modified tests. For example:
     # REPEATED_UTESTS_LONG: org.apache.cassandra.db.commitlog.CommitLogStressTest
@@ -537,6 +544,30 @@ j11_separate_jobs: &j11_separate_jobs
         requires:
           - start_j17_utests_fqltool_repeat
           - j11_build
+    - start_j11_utests_sstableloader:
+        type: approval
+    - j11_utests_sstableloader:
+        requires:
+          - start_j11_utests_sstableloader
+          - j11_build
+    - start_j17_utests_sstableloader:
+        type: approval
+    - j17_utests_sstableloader:
+        requires:
+          - start_j17_utests_sstableloader
+          - j11_build
+    - start_j11_utests_sstableloader_repeat:
+        type: approval
+    - j11_utests_sstableloader_repeat:
+        requires:
+          - start_j11_utests_sstableloader_repeat
+          - j11_build
+    - start_j17_utests_sstableloader_repeat:
+        type: approval
+    - j17_utests_sstableloader_repeat:
+        requires:
+          - start_j17_utests_sstableloader_repeat
+          - j11_build
     - start_j11_utests_system_keyspace_directory:
         type: approval
     - j11_utests_system_keyspace_directory:
@@ -974,6 +1005,24 @@ j11_pre-commit_jobs: &j11_pre-commit_jobs
         requires:
           - start_utests_fqltool
           - j11_build
+    - start_utests_sstableloader:
+        type: approval
+    - j11_utests_sstableloader:
+        requires:
+          - start_utests_sstableloader
+          - j11_build
+    - j17_utests_sstableloader:
+        requires:
+          - start_utests_sstableloader
+          - j11_build
+    - j11_utests_sstableloader_repeat:
+        requires:
+          - start_utests_sstableloader
+          - j11_build
+    - j17_utests_sstableloader_repeat:
+        requires:
+          - start_utests_sstableloader
+          - j11_build
     - start_utests_system_keyspace_directory:
         type: approval
     - j11_utests_system_keyspace_directory:
@@ -1356,6 +1405,18 @@ j17_separate_jobs: &j17_separate_jobs
         requires:
           - start_j17_utests_fqltool_repeat
           - j17_build
+    - start_j17_utests_sstableloader:
+        type: approval
+    - j17_utests_sstableloader:
+        requires:
+          - start_j17_utests_sstableloader
+          - j17_build
+    - start_j17_utests_sstableloader_repeat:
+        type: approval
+    - j17_utests_sstableloader_repeat:
+        requires:
+          - start_j17_utests_sstableloader_repeat
+          - j17_build
     - start_j17_utests_system_keyspace_directory:
         type: approval
     - j17_utests_system_keyspace_directory:
@@ -1540,6 +1601,16 @@ j17_pre-commit_jobs: &j17_pre-commit_jobs
     - j17_utests_fqltool_repeat:
         requires:
           - start_utests_fqltool
+          - j17_build
+    - start_utests_sstableloader:
+        type: approval
+    - j17_utests_sstableloader:
+        requires:
+          - start_utests_sstableloader
+          - j17_build
+    - j17_utests_sstableloader_repeat:
+        requires:
+          - start_utests_sstableloader
           - j17_build
     - start_utests_system_keyspace_directory:
         type: approval
@@ -1874,6 +1945,22 @@ jobs:
           at: /home/cassandra
       - run_junit_tests:
           target: fqltool-test
+
+  j11_utests_sstableloader:
+    <<: *j11_seq_executor
+    steps:
+      - attach_workspace:
+          at: /home/cassandra
+      - run_junit_tests:
+          target: sstableloader-test
+
+  j17_utests_sstableloader:
+    <<: *j17_seq_executor
+    steps:
+      - attach_workspace:
+          at: /home/cassandra
+      - run_junit_tests:
+          target: sstableloader-test
 
   j11_utests_system_keyspace_directory:
     <<: *j11_par_executor
@@ -2379,6 +2466,22 @@ jobs:
           at: /home/cassandra
       - log_environment
       - run_utests_fqltool_repeat
+
+  j11_utests_sstableloader_repeat:
+    <<: *j11_repeated_utest_executor
+    steps:
+      - attach_workspace:
+          at: /home/cassandra
+      - log_environment
+      - run_utests_sstableloader_repeat
+
+  j17_utests_sstableloader_repeat:
+    <<: *j17_repeated_utest_executor
+    steps:
+      - attach_workspace:
+          at: /home/cassandra
+      - log_environment
+      - run_utests_sstableloader_repeat
 
   j11_utests_long_repeat:
     <<: *j11_repeated_utest_executor
@@ -3114,6 +3217,14 @@ commands:
           count: ${REPEATED_UTESTS_FQLTOOL_COUNT}
           stop_on_failure: ${REPEATED_TESTS_STOP_ON_FAILURE}
 
+  run_utests_sstableloader_repeat:
+    steps:
+      - run_repeated_utests:
+          target: sstableloader-test
+          tests: ${REPEATED_UTESTS_SSTABLELOADER}
+          count: ${REPEATED_UTESTS_SSTABLELOADER_COUNT}
+          stop_on_failure: ${REPEATED_TESTS_STOP_ON_FAILURE}
+
   run_utests_stress_repeat:
     steps:
       - run_repeated_utests:
@@ -3237,6 +3348,7 @@ commands:
                       $target == "test-oa" || \
                       $target == "test-system-keyspace-directory" || \
                       $target == "fqltool-test" || \
+                      $target == "sstableloader-test" || \
                       $target == "long-test" || \
                       $target == "stress-test" || \
                       $target == "test-simulator-dtest" ]]; then
@@ -3365,6 +3477,7 @@ commands:
                       $target == "test-oa" || \
                       $target == "test-system-keyspace-directory" || \
                       $target == "fqltool-test" || \
+                      $target == "sstableloader-test" || \
                       $target == "long-test" || \
                       $target == "stress-test" || \
                       $target == "test-simulator-dtest" ]]; then
