+++ ide/idea-iml-file.xml
@@ -30,6 +30,8 @@
             <sourceFolder url="file://$MODULE_DIR$/tools/stress/test/unit" isTestSource="true" />
             <sourceFolder url="file://$MODULE_DIR$/tools/fqltool/src" isTestSource="false" />
             <sourceFolder url="file://$MODULE_DIR$/tools/fqltool/test/unit" isTestSource="true" />
+            <sourceFolder url="file://$MODULE_DIR$/tools/sstableloader/src" isTestSource="false" />
+            <sourceFolder url="file://$MODULE_DIR$/tools/sstableloader/test/unit" isTestSource="true" />
             <sourceFolder url="file://$MODULE_DIR$/test/unit" isTestSource="true" />
             <sourceFolder url="file://$MODULE_DIR$/test/long" isTestSource="true" />
             <sourceFolder url="file://$MODULE_DIR$/test/memory" isTestSource="true" />
