+++ ide/idea/workspace.xml
@@ -360,6 +360,7 @@
         <filter targetName="build-project" isVisible="false" />
         <filter targetName="stress-build" isVisible="true" />
         <filter targetName="fqltool-build" isVisible="true" />
+        <filter targetName="sstableloader-build" isVisible="true" />
         <filter targetName="_write-poms" isVisible="false" />
         <filter targetName="write-poms" isVisible="false" />
         <filter targetName="jar" isVisible="true" />
