+++ src/java/org/apache/cassandra/tools/StandaloneSSTableUtil.java
@@ -30,7 +30,6 @@ import java.io.IOException;
 import java.util.function.BiPredicate;
 
 import org.apache.cassandra.io.util.File;
-import static org.apache.cassandra.tools.BulkLoader.CmdLineOptions;
 
 public class StandaloneSSTableUtil
 {
