+++ tools/sstableloader/src/org/apache/cassandra/tools/LoaderOptions.java
@@ -49,7 +49,6 @@ import org.apache.cassandra.config.YamlConfigurationLoader;
 import org.apache.cassandra.exceptions.ConfigurationException;
 import org.apache.cassandra.io.util.File;
 import org.apache.cassandra.locator.InetAddressAndPort;
-import org.apache.cassandra.tools.BulkLoader.CmdLineOptions;
 
 import static org.apache.cassandra.config.DataRateSpec.DataRateUnit.MEBIBYTES_PER_SECOND;
 import static org.apache.cassandra.config.EncryptionOptions.ClientEncryptionOptions.ClientAuth.REQUIRED;
@@ -77,9 +76,8 @@ public class LoaderOptions
     /**
      * Throttle defined in megabits per second. CASSANDRA-10637 introduced a builder and is the preferred way to
      * provide options instead of using these constant fields.
-     * @deprecated Use {@code throttle-mib} instead
+     * @deprecated Use {@code throttle-mib} instead. See CASSANDRA-17677
      */
-    /** @deprecated See CASSANDRA-17677 */
     @Deprecated(since = "5.0")
     public static final String THROTTLE_MBITS = "throttle";
     public static final String THROTTLE_MEBIBYTES = "throttle-mib";
