+++ tools/sstableloader/src/org/apache/cassandra/utils/NativeSSTableLoaderClient.java
